backend/requirements.txt
Flask
Tensorflow
Pillow
Numpy
Flask-Cors
